requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.31.0
attrs==26.1.0
audioread==3.1.0
babel==2.18.0
backports.zstd==1.3.0
bangla==0.0.5
bcrypt==5.0.0
black==26.3.1
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
brotli==1.2.0
build==1.4.2
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
confection==1.3.3
contourpy==1.2.1
coqpit==0.0.17
coqpit-config==0.2.4
coqui-tts-trainer==0.3.3
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
discord.py==2.7.1
docopt==0.6.2
durationpy==0.10
einops==0.8.2
encodec==0.1.1
exceptiongroup==1.3.1
filelock==3.25.2
Flask==3.1.3
flask-cloudflared==0.0.15
Flask-Compress==1.24
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
fugashi==1.5.2
g2pkk==0.1.2
googleapis-common-protos==1.74.0
grpcio==1.80.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflect==7.5.0
itsdangerous==2.2.0
jaconv==0.5.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
jsonlines==1.2.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
ko-speech-tools==0.1.0
kubernetes==35.0.0
lazy-loader==0.5
librosa==0.10.0
llvmlite==0.47.0
loguru==0.7.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
mdurl==0.1.2
mecab-python3==1.0.6
mmh3==5.2.1
mojimoji==0.0.13
monotonic-alignment-search==0.2.1
more-itertools==11.0.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==2.8.8
nltk==3.9.4
num2words==0.5.14
numba==0.65.0
numpy==1.22.0
oauthlib==3.3.1
onnxruntime==1.23.2
openai-whisper==20250625
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
outcome==1.3.0.post0
overrides==7.7.0
packaging==26.0
pandas==1.5.3
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
preshed==3.0.13
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
py-espeak-ng==0.1.8
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.20.0
pynndescent==0.6.0
pyparsing==3.3.2
PyPika==0.51.1
pypinyin==0.55.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.11.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
srt==3.5.3
starlette==1.0.0
SudachiDict-core==20260116
SudachiPy==0.6.10
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.15.2
tomli==2.4.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
tqdm==4.67.3
trainer==0.0.36
transformers==4.36.2
trio==0.33.0
trio-websocket==0.12.2
TTS==0.22.0
typeguard==4.5.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.4.0
unidic-lite==1.0.8
urllib3==2.6.3
uvicorn==0.43.0
vosk==0.3.45
wasabi==1.1.3
watchfiles==1.1.1
weasel==1.0.0
websocket-client==1.9.0
websockets==16.0
webuiapi==0.9.17
Werkzeug==3.1.8
win32_setctime==1.2.0
wrapt==2.1.2
wsproto==1.3.2
wxPython==4.2.5
yarl==1.23.0
zipp==3.23.0
